# nestjs-backend/swagger-doc.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/asset": {
      "get": {
        "operationId": "AssetController_getAll",
        "summary": "Get all Assets",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of all assets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "example": [
                      {
                        "_id": "1",
                        "ticker": "AMGN",
                        "name": "Amgen",
                        "type": "STOCK",
                        "geography": "USA",
                        "industry": "Biopharmaceutical"
                      },
                      {
                        "_id": "2",
                        "ticker": "DOW",
                        "name": "Dow",
                        "type": "STOCK",
                        "geography": "USA",
                        "industry": "Chemical"
                      },
                      {
                        "_id": "3",
                        "ticker": "AAPL",
                        "name": "Apple",
                        "type": "STOCK",
                        "geography": "USA",
                        "industry": "Technology"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Asset Service"
        ]
      },
      "post": {
        "operationId": "AssetController_create",
        "summary": "Create an Asset",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Details of the asset to create",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetDto"
              },
              "examples": {
                "example1": {
                  "summary": "Example Asset",
                  "value": {
                    "ticker": "GOOGL",
                    "name": "Alphabet Inc.",
                    "sector": "Technology",
                    "cost": 2800.5
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Asset created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "_id": "1",
                    "ticker": "AAPL",
                    "name": "Apple",
                    "type": "STOCK",
                    "geography": "USA",
                    "industry": "Technology"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Asset Service"
        ]
      }
    },
    "/asset/{ticker}": {
      "get": {
        "operationId": "AssetController_getByTicker",
        "summary": "Get one Asset by ticker",
        "parameters": [
          {
            "name": "ticker",
            "required": true,
            "in": "path",
            "description": "The ticker symbol of the asset",
            "example": "AAPL",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Asset details by ticker",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "_id": "3",
                    "ticker": "AAPL",
                    "name": "Apple",
                    "type": "STOCK",
                    "geography": "USA",
                    "industry": "Technology"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Asset Service"
        ]
      },
      "put": {
        "operationId": "AssetController_update",
        "summary": "Update an Asset by ticker",
        "parameters": [
          {
            "name": "ticker",
            "required": true,
            "in": "path",
            "description": "The ticker symbol of the asset to update",
            "example": "MSFT",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Updated asset details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetDto"
              },
              "examples": {
                "example1": {
                  "summary": "Updated Asset Example",
                  "value": {
                    "ticker": "MSFT",
                    "name": "Microsoft Corporation",
                    "type": "STOCK",
                    "geography": "USA",
                    "industry": "Technology"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Asset updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "_id": "1",
                    "ticker": "AAPL",
                    "name": "Apple",
                    "type": "STOCK",
                    "geography": "USA",
                    "industry": "Technology"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Asset Service"
        ]
      },
      "delete": {
        "operationId": "AssetController_delete",
        "summary": "Delete an Asset by ticker",
        "parameters": [
          {
            "name": "ticker",
            "required": true,
            "in": "path",
            "description": "The ticker symbol of the asset to delete",
            "example": "JPM",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "204": {
            "description": "Asset deleted successfully"
          }
        },
        "tags": [
          "Asset Service"
        ]
      }
    },
...
  },
  "info": {
    "title": "Portfolio Optimizer Application API",
    "description": "core application in nestjs",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "AssetDto": {
        "type": "object",
        "properties": {
          "ticker": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "enum": [
              "STOCK",
              "BOND"
            ],
            "type": "string"
          },
          "geography": {
            "type": "string"
          },
          "industry": {
            "type": "string"
          }
        },
        "required": [
          "ticker",
          "name",
          "type",
          "geography",
          "industry"
        ]
      },
...
    }
  }
}